Refactoring generate and parse
